renamed cucumberjs to cucumber
